requirements.txt
asgiref==3.9.1
attrs==25.3.0
braintree==4.38.0
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
pillow==11.3.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
